requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
jinja2==3.1.5
python-multipart==0.0.20
slowapi==0.1.9
kubernetes==31.0.0
